refactor(desktop): question telemetry lives in its own extension file
